requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1617740417847/work
certifi==2020.12.5
cycler==0.10.0
defusedxml==0.7.1
diff-match-patch==20200713
Django @ file:///tmp/build/80754af9/django_1617740316746/work
django-import-export==2.5.0
et-xmlfile==1.1.0
kiwisolver==1.3.1
MarkupPy==1.14
matplotlib==3.4.2
numpy==1.20.3
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
psycopg2 @ file:///C:/ci/psycopg2_1612298803839/work
pyparsing==2.4.7
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.4.1
six==1.16.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tablib==3.0.0
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
